requirements.txt
tabula-py
pdfminer.six